requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
jinja2==3.1.3
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0